requirements.txt
openai==1.45.0
yaspin
python-docx
python-dotenv

